Move event key code map to `keyCodes.ts`
